vnese-id-be-python/requirements.txt
# FastAPI và các thành phần web
fastapi>=0.103.1
uvicorn>=0.23.2
jinja2>=3.1.2
python-multipart>=0.0.6
websockets>=11.0.3

# Xử lý dữ liệu và cơ sở dữ liệu
sqlalchemy>=2.0.21
pydantic>=2.4.2
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.7
alembic>=1.12.0

# Bảo mật và xác thực (nếu cần dùng về sau)
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Môi trường và cấu hình
python-dotenv>=1.0.0

# Thư viện cho deep learning và YOLOv5
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.13.0
PyYAML>=6.0
tqdm>=4.66.0

# Thư viện testing
pytest>=7.4.2
httpx>=0.25.0

# Các thư viện tùy chọn cho YOLOv5
# Bỏ comment nếu cần tích hợp
# wandb
# clearml>=1.10.0
# ultralytics>=8.0.0 